lessons/lesson_19_rest_api/requirements.txt
requests
pytest
voluptuous
pytest-voluptuous
allure-pytest
curlify
selene==2.0.0rc2
